requirements.txt
# SenseFlow Requirements
# 
# Note: This codebase uses CUDA 12.4 or compatible.
# Before installing these requirements, please install PyTorch with CUDA support:
#   pip install torch==2.5.1 torchvision==0.20.1 torchaudio==2.5.1 --index-url https://download.pytorch.org/whl/cu124
#
# The NVIDIA CUDA packages (nvidia-cublas-cu12, nvidia-cudnn-cu12, etc.) will be
# automatically installed as dependencies when installing PyTorch with CUDA support.
# They are not listed here to allow flexibility for different CUDA versions.

GitPython==3.1.43
Jinja2==3.1.5
MarkupSafe==3.0.2
PyYAML==6.0.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==24.3.0
autocommand==2.2.2
bitsandbytes==0.48.2
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
cryptography==46.0.3
diffusers==0.32.1
docstring_parser==0.17.0
einops==0.8.0
filelock==3.16.1
ftfy==6.3.1
future==1.0.0
gitdb==4.0.11
huggingface-hub==0.36.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.0
lmdb==1.7.5
more-itertools==10.3.0
ninja==1.11.1.3
numpy==2.2.1
omegaconf==2.3.0
open_clip_torch==2.31.0
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.2
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
six==1.17.0
timm==1.0.15
tokenizers==0.21.4
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.50.0
triton==3.1.0
typing_extensions==4.15.0
urllib3==2.3.0
-e .
